+++ lib/karafka/connection/listener.rb
@@ -267,7 +267,7 @@ module Karafka
           # Start work coordination for this topic partition
           coordinator.start(messages)
 
-          @partitioner.call(topic, messages) do |group_id, partition_messages|
+          @partitioner.call(topic, messages, coordinator) do |group_id, partition_messages|
             # Count the job we're going to create here
             coordinator.increment
             executor = @executors.find_or_create(topic, partition, group_id)