+++ spec/integrations/consumption/strategies/aj_mom/backoff_and_retry_on_error_spec.rb
@@ -15,7 +15,7 @@ class Job < ActiveJob::Base
   queue_as DT.topic
 
   def perform
-    if DT[0].size.zero?
+    if DT[0].empty?
       DT[0] << '1'
       raise StandardError
     else